move values in store
