+++ hw06-ATM_manager/src/main/java/ATM/Operation/Operation.java
@@ -4,6 +4,8 @@
 
 package ATM.Operation;
 
+import ATM.Store.Value;
+
 public interface Operation {
     int action(int needSum, int countBills, Value denomination);
 }
