+++ hw06-ATM_manager/src/main/java/ATM/Store/Store.java
@@ -10,4 +10,5 @@ public interface Store {
     void putBills(int count);
     int getCountBills();
     int getSumCountBills();
+    Value getDenomination();
 }
