+++ hw06-ATM_manager/src/main/java/ATM/Store/Value.java
@@ -1,4 +1,4 @@
-package ATM.Operation;
+package ATM.Store;
 
 /**
  * Created by Ilya Rogatkin, May 2020
